requirements.txt
# Core
streamlit>=1.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM
llama-cpp-python>=0.2.0

# LangGraph Agent
langgraph>=0.0.40

# Embeddings & Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Reranking
flashrank>=0.2.0

# PDF Processing
pypdf>=3.0.0

# Text Splitting
langchain-text-splitters>=0.0.1

# Knowledge Graph
# (uses stdlib sqlite3)

# Logging
structlog>=23.0.0

# Translation (BUET BanglaT5)
transformers>=4.30.0
sentencepiece>=0.1.99
bnunicodenormalizer>=0.1.5

# Voice Input (Whisper ASR)
faster-whisper>=1.0.0

# Voice Output (TTS)
pyttsx3>=2.90

# Audio Recording/Processing
sounddevice>=0.4.6
soundfile>=0.12.0
